move examples to top-level directory
