+++ docs/content/examples/overcooked_multiplayer.rst
@@ -76,7 +76,7 @@ From the repository root, run as a module:
 
 .. code-block:: bash
 
-    python -m mug.examples.cogrid.overcooked_human_human_server_side
+    python -m examples.cogrid.overcooked_human_human_server_side
 
 Then:
 
@@ -112,7 +112,7 @@ The experiment uses a Stager to manage scene progression for both players:
 .. code-block:: python
 
     from mug.scenes import stager, scene
-    from mug.examples.cogrid.scenes import scenes as oc_scenes
+    from examples.cogrid.scenes import scenes as oc_scenes
 
     stager = stager.Stager(
         scenes=[
