+++ examples/cogrid/assets/alt/closed_pot.png
