+++ examples/cogrid/assets/alt/empty_bowl.png
