+++ examples/cogrid/assets/alt/empty_pot.png
