+++ examples/cogrid/assets/alt/full_bowl.png
