+++ examples/cogrid/assets/alt/full_pot.png
