+++ examples/cogrid/assets/alt/onion.png
