+++ examples/cogrid/assets/objects.json
