+++ examples/cogrid/assets/objects.png
