+++ examples/cogrid/assets/soups.json
