+++ examples/cogrid/assets/terrain.json
