+++ examples/cogrid/assets/terrain.png
