+++ examples/cogrid/assets/test_traj.json
