+++ examples/cogrid/assets/tiles.png
