+++ examples/cogrid/environments/cramped_room_environment_initialization_hh.py
