+++ examples/cogrid/environments/tutorial_cramped_room_environment_initialization.py
