+++ examples/cogrid/html_pages/choice_cramped_room.html
