+++ examples/cogrid/html_pages/fixed_cramped_room.html
