+++ examples/footsies/footsies_scene.py
