+++ examples/footsies/static/controllable_difficulty.html
