+++ examples/footsies/static/tutorial_static.html
