+++ examples/mountain_car/mountain_car_experiment.py
@@ -74,7 +74,7 @@ mountain_car_scene = (
     )
     .runtime(
         run_through_pyodide=True,
-        environment_initialization_code_filepath="mug/examples/mountain_car/mountain_car_rgb_env.py",
+        environment_initialization_code_filepath="examples/mountain_car/mountain_car_rgb_env.py",
     )
 )
