+++ examples/mountain_car/mountain_car_rgb_env.py
