+++ tests/fixtures/overcooked_human_ai_test.py
@@ -22,8 +22,8 @@ eventlet.monkey_patch()
 import argparse
 import copy
 
+from examples.cogrid.scenes import scenes as oc_scenes
 from mug.configurations import experiment_config
-from mug.examples.cogrid.scenes import scenes as oc_scenes
 from mug.scenes import stager
 from mug.server import app
