+++ tests/fixtures/overcooked_human_human_multiplayer_multi_episode_test.py
@@ -26,8 +26,8 @@ eventlet.monkey_patch()
 
 import argparse
 
+from examples.cogrid.scenes import scenes as oc_scenes
 from mug.configurations import experiment_config
-from mug.examples.cogrid.scenes import scenes as oc_scenes
 from mug.scenes import stager, static_scene
 from mug.server import app
 from mug.server.matchmaker import FIFOMatchmaker
